backend/requirements.txt
dlib==19.24.6
dnspython==2.7.0
numpy==2.2.3
opencv-python==4.11.0.86
pymongo==4.11.1
Flask==3.0.3
Flask-Cors==4.0.1
python-dotenv==1.0.1
SQLAlchemy==2.0.32
alembic==1.13.2
opencv-python==4.10.0.84
numpy==1.26.4
dlib==19.24.6
cmake
